+++ lib/src/helper/action/shortcut_action.dart
@@ -1,11 +1,1 @@
-/* 
-
-              Copyright (c) 2021 divshekhar (Divyanshu Shekhar). 
-                            All rights reserved.
-
-The plugin is governed by the BSD-3-clause License. Please see the LICENSE file
-for more details.
-
-*/
-
-typedef void ShortcutAction(String action);
+typedef ShortcutAction = void Function(String action);