+++ lib/src/helper/enums/shortcut_icon_asset.dart
@@ -1,13 +1,3 @@
-/* 
-
-              Copyright (c) 2021 divshekhar (Divyanshu Shekhar). 
-                            All rights reserved.
-
-The plugin is governed by the BSD-3-clause License. Please see the LICENSE file
-for more details.
-
-*/
-
 enum ShortcutIconAsset {
   /// Creates Icon from the native resources `drawable` & `mipmap` directory.
   nativeAsset,