+++ lib/src/helper/helper.dart
@@ -1,13 +1,3 @@
-/* 
-
-              Copyright (c) 2021 divshekhar (Divyanshu Shekhar). 
-                            All rights reserved.
-
-The plugin is governed by the BSD-3-clause License. Please see the LICENSE file
-for more details.
-
-*/
-
 export 'action/shortcut_action.dart';
-export 'model/shortcut_item.dart';
 export 'enums/shortcut_icon_asset.dart';
+export 'model/shortcut_item.dart';