+++ lib/src/helper/model/shortcut_item.dart
@@ -1,13 +1,3 @@
-/* 
-
-              Copyright (c) 2021 divshekhar (Divyanshu Shekhar). 
-                            All rights reserved.
-
-The plugin is governed by the BSD-3-clause License. Please see the LICENSE file
-for more details.
-
-*/
-
 import '../enums/shortcut_icon_asset.dart';
 
 class ShortcutItem {